I forget to move all the dto classes
